+++ src/core/CCMap.mli
@@ -16,6 +16,7 @@ module type OrderedType = Map.OrderedType
 
 module type S = sig
   include Map.S
+  (** @inline *)
 
   val get : key -> 'a t -> 'a option
   (** [get k m] returns [Some v] if the current binding of [k] in [m] is [v],
@@ -27,47 +28,6 @@ module type S = sig
       and returns [default] otherwise (if [k] doesn't belong in [m]).
       @since 0.16 *)
 
-  val update : key -> ('a option -> 'a option) -> 'a t -> 'a t
-  (** [update k f m] calls [f (Some v)] if [find k m = v],
-      otherwise it calls [f None]. In any case, if the result is [None]
-      [k] is removed from [m], and if the result is [Some v'] then
-      [add k v' m] is returned. *)
-
-  val choose_opt : 'a t -> (key * 'a) option
-  (** [choose_opt m] returns one binding of the given map [m], or [None] if [m] is empty.
-      Safe version of {!choose}.
-      @since 1.5 *)
-
-  val min_binding_opt : 'a t -> (key * 'a) option
-  (** [min_binding_opt m] returns the smallest binding of the given map [m],
-      or [None] if [m] is empty.
-      Safe version of {!min_binding}.
-      @since 1.5 *)
-
-  val max_binding_opt : 'a t -> (key * 'a) option
-  (** [max_binding_opt m] returns the largest binding of the given map [m],
-      or [None] if [m] is empty.
-      Safe version of {!max_binding}.
-      @since 1.5 *)
-
-  val find_opt : key -> 'a t -> 'a option
-  (** [find_opt k m] returns [Some v] if the current binding of [k] in [m] is [v],
-      or [None] if the key [k] is not present.
-      Safe version of {!find}.
-      @since 1.5 *)
-
-  val find_first : (key -> bool) -> 'a t -> key * 'a
-  (** [find_first f m] where [f] is a monotonically increasing function, returns the binding of [m]
-      with the lowest key [k] such that [f k], or raises [Not_found] if no such key exists.
-      See {!Map.S.find_first}.
-      @since 1.5 *)
-
-  val find_first_opt : (key -> bool) -> 'a t -> (key * 'a) option
-  (** [find_first_opt f m] where [f] is a monotonically increasing function, returns an option containing
-      the binding of [m] with the lowest key [k] such that [f k], or [None] if no such key exists.
-      Safe version of {!find_first}.
-      @since 1.5 *)
-
   val merge_safe :
     f:(key -> [ `Left of 'a | `Right of 'b | `Both of 'a * 'b ] -> 'c option) ->
     'a t ->
@@ -75,12 +35,6 @@ module type S = sig
     'c t
   (** [merge_safe ~f a b] merges the maps [a] and [b] together.
       @since 0.17 *)
-
-  val add_seq : 'a t -> (key * 'a) Seq.t -> 'a t
-  (** [add_seq m seq] adds the given [Seq.t] of bindings to the map [m].
-      Like {!add_list}.
-      Renamed from [add_std_seq] since 3.0.
-      @since 3.0 *)
 
   val add_seq_with :
     f:(key -> 'a -> 'a -> 'a) -> 'a t -> (key * 'a) Seq.t -> 'a t
@@ -90,12 +44,6 @@ module type S = sig
       function [f], with [f key v1 v2] being called with [v1] occurring
       later in the seq than [v2].
       @since 3.3 *)
-
-  val of_seq : (key * 'a) Seq.t -> 'a t
-  (** [of_seq seq] builds a map from the given [Seq.t] of bindings.
-      Like {!of_list}.
-      Renamed from [of_std_seq] since 3.0.
-      @since 3.0 *)
 
   val of_seq_with : f:(key -> 'a -> 'a -> 'a) -> (key * 'a) Seq.t -> 'a t
   (** [of_seq_with ~f l] builds a map from the given seq [l] of bindings [k_i -> v_i],