+++ src/core/CCResult.mli
@@ -16,13 +16,8 @@ type 'a printer = Format.formatter -> 'a -> unit
 
 (** {2 Basics} *)
 
-type nonrec (+'good, +'bad) result = ('good, 'bad) result =
-  | Ok of 'good
-  | Error of 'bad
-
-type (+'good, +'bad) t = ('good, 'bad) result =
-  | Ok of 'good
-  | Error of 'bad
+include module type of Result
+(** @inline *)
 
 val return : 'a -> ('a, 'err) t
 (** Successfully return a value. *)
@@ -68,22 +63,15 @@ val opt_map : ('a -> ('b, 'c) t) -> 'a option -> ('b option, 'c) t
 (** Map a fallible operation through an option.
     @since 3.7 *)
 
-val map : ('a -> 'b) -> ('a, 'err) t -> ('b, 'err) t
-(** Map on success. *)
-
 val map_err : ('err1 -> 'err2) -> ('a, 'err1) t -> ('a, 'err2) t
-(** Map on the error variant. *)
+(** Alias of [map_error] *)
 
 val map2 : ('a -> 'b) -> ('err1 -> 'err2) -> ('a, 'err1) t -> ('b, 'err2) t
 (** Like {!map}, but also with a function that can transform
     the error message in case of failure. *)
 
-val iter : ('a -> unit) -> ('a, _) t -> unit
-(** Apply the function only in case of [Ok]. *)
-
 val iter_err : ('err -> unit) -> (_, 'err) t -> unit
-(** Apply the function in case of [Error].
-    @since 2.4 *)
+(** Alias of {!iter_error} *)
 
 exception Get_error
 
@@ -120,6 +108,13 @@ val catch : ('a, 'err) t -> ok:('a -> 'b) -> err:('err -> 'b) -> 'b
 
 val flat_map : ('a -> ('b, 'err) t) -> ('a, 'err) t -> ('b, 'err) t
 
+[@@@iflt 5.4]
+
+val retract : ('a, 'a) t -> 'a
+(** [retract r] collapse [r] to [v] if [r] is either [Ok v] or [Error v]. *)
+
+[@@@endif]
+
 val k_compose :
   ('a -> ('b, 'err) t) -> ('b -> ('c, 'err) t) -> 'a -> ('c, 'err) t
 (** Kleisli composition. Monadic equivalent of {!CCFun.compose}.
@@ -128,22 +123,10 @@ val k_compose :
 val equal : err:'err equal -> 'a equal -> ('a, 'err) t equal
 val compare : err:'err ord -> 'a ord -> ('a, 'err) t ord
 
-val fold : ok:('a -> 'b) -> error:('err -> 'b) -> ('a, 'err) t -> 'b
-(** [fold ~ok ~error e] opens [e] and, if [e = Ok x], returns
-    [ok x], otherwise [e = Error s] and it returns [error s]. *)
-
 val fold_ok : ('a -> 'b -> 'a) -> 'a -> ('b, _) t -> 'a
 (** [fold_ok f acc r] will compute [f acc x] if [r=Ok x],
     and return [acc] otherwise, as if the result were a mere option.
     @since 1.2 *)
-
-val is_ok : ('a, 'err) t -> bool
-(** Return true if [Ok].
-    @since 1.0 *)
-
-val is_error : ('a, 'err) t -> bool
-(** Return true if [Error].
-    @since 1.0 *)
 
 (** {2 Wrappers} *)
 
@@ -172,14 +155,17 @@ val wrap3 : ('a -> 'b -> 'c -> 'd) -> 'a -> 'b -> 'c -> ('d, exn) t
 val pure : 'a -> ('a, 'err) t
 (** Synonym of {!return}. *)
 
-val join : (('a, 'err) t, 'err) t -> ('a, 'err) t
-(** [join t], in case of success, returns [Ok o] from [Ok (Ok o)]. Otherwise,
-    it fails with [Error e] where [e] is the unwrapped error of [t]. *)
+[@@@iflt 5.4]
 
-val both : ('a, 'err) t -> ('b, 'err) t -> ('a * 'b, 'err) t
-(** [both a b], in case of success, returns [Ok (o, o')] with the ok values
+val product : ('a, 'err) t -> ('b, 'err) t -> ('a * 'b, 'err) t
+(** [product a b], in case of success, returns [Ok (o, o')] with the ok values
     of [a] and [b]. Otherwise, it fails, and the error of [a] is chosen over the
     error of [b] if both fail. *)
+
+[@@@endif]
+
+val both : ('a, 'err) t -> ('b, 'err) t -> ('a * 'b, 'err) t
+(** Alias of {!product} *)
 
 (** {2 Infix} *)
 
@@ -279,17 +265,13 @@ end
 (** {2 Conversions} *)
 
 val to_opt : ('a, _) t -> 'a option
-(** Convert a result to an option. *)
+(** Alias of {!to_option} *)
 
 val of_opt : 'a option -> ('a, string) t
 (** [of_opt opt] converts [Some v] to [Ok v] and [None] to [Error "of_opt"].*)
 
 val to_iter : ('a, _) t -> 'a iter
 (** @since 2.8 *)
-
-val to_seq : ('a, _) t -> 'a Seq.t
-(** Renamed from [to_std_seq] since 3.0.
-    @since 3.0 *)
 
 type ('a, 'b) error =
   [ `Ok of 'a