+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+## main
+- breaking: CCListLabel.compare and CCListLabel.equal takes the function on the elements as named arguments
+- breaking: CCListLabel.init now takes the length as a named arguments to follow the Stdlib
+- breaking: change the semantic of CCFloat.{min,max} with respect to NaN to follow the Stdlib
+- breaking: change the semantic of CCInt.rem with respect to negative number to follow the Stdlib
+- breaking: change the order of argument of CCMap.add_seq to align with the stdlib.
 
 ## 3.17
 
@@ -60,6 +66,7 @@
 
 ## 3.13
 
+- breaking: bump minimum version of OCaml to 4.08
 - breaking: delete containers-thread (which was deprecated)
 - breaking: pp: modify `Ext.t` so it takes surrounding value
 - breaking: remove CCShims